Model/requirement.txt
numpy
pytorch-gpu
scipy
networkx
scikit-learn
umap-learn
dill
